requirements.txt
streamlit
python-docx
google-generativeai
docx2pdf     # optional; requires MS Word on machine (Windows/Mac)
python-dotenv